Task/4.15 - 4.21

+++ osa4/controllers/blogs.js
@@ -1,10 +1,14 @@
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({});
+    const blogs = await Blog
+      .find({})
+      .populate('user', { username: 1, name: 1, id: 1 });
     response.json(blogs.map(blog => blog.toJSON()))
   } catch(e) {
     next(e);
@@ -12,10 +16,24 @@ blogsRouter.get('/', async (request, response, next) => {
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+  const blogBase = { ...request.body };
   try {
-    const result = await blog.save();
-    response.status(201).json(result);
+    const decodedToken = request.token
+      ? jwt.verify(request.token, process.env.SECRET)
+      : null;
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const userToUpdate = await User.findById(decodedToken.id);
+
+    blogBase.user = userToUpdate._id;
+    const blog = new Blog(blogBase)
+    const savedBlog = await blog.save();
+
+    userToUpdate.blogs.push(savedBlog._id);
+    await userToUpdate.save();
+
+    response.status(201).json(savedBlog);
   } catch(e) {
     next(e);
   }
@@ -23,6 +41,16 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
   try {
+    const decodedToken = req.token
+      ? jwt.verify(req.token, process.env.SECRET)
+      : null;
+    if (!req.token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blogToDelete = await Blog.findById(req.params.id, 'user');
+    if (decodedToken.id !== blogToDelete.user.toString()) {
+      return res.status(401).json({ error: 'token invalid' })
+    }
     await Blog.findByIdAndRemove(req.params.id);
     res.status(204).end();
   } catch(exception) {
